[PF-127] Move enums/autovalue classes to common place (#17)
* update

* split steps

* split steps

* split steps

* add test dir

* add test dir

* add test dir

* add test dir

* add test dir

* address comment

* address comment

* delete flightManager test

* tmp

* add one more test

* add one more test

* address comment

* address comment

* address comments

* address comments

* more polling times since flight make take long

* more polling times since flight make take long

* move enums/autovalue classes to common
