+++ src/main/java/bio/terra/rbs/common/RequestHandoutId.java
@@ -1,8 +1,8 @@
-package bio.terra.rbs.db;
+package bio.terra.rbs.common;
 
 import com.google.auto.value.AutoValue;
 
-/** Wraps the request_handout_id in db resource table. */
+/** Wraps the request_handout_id used in RBS. */
 @AutoValue
 public abstract class RequestHandoutId {
   public abstract String id();
