+++ src/main/java/bio/terra/rbs/common/Resource.java
@@ -1,11 +1,11 @@
-package bio.terra.rbs.db;
+package bio.terra.rbs.common;
 
 import bio.terra.rbs.generated.model.CloudResourceUid;
 import com.google.auto.value.AutoValue;
 import java.time.Instant;
 import javax.annotation.Nullable;
 
-/** Represents a record in the resource table in the RBS database. */
+/** Represents resource used in RBS.. */
 @AutoValue
 public abstract class Resource {
   /** The unique id of the resources. Generated by RBS when creating the resource. */
