+++ src/main/java/bio/terra/rbs/service/resource/flight/FinishResourceCreationStep.java
@@ -1,5 +1,6 @@
 package bio.terra.rbs.service.resource.flight;
 
+import bio.terra.rbs.common.ResourceId;
 import bio.terra.rbs.db.*;
 import bio.terra.rbs.generated.model.CloudResourceUid;
 import bio.terra.rbs.service.resource.FlightMapKeys;
