+++ src/test/java/bio/terra/rbs/service/resource/FlightFactoryImplTest.java
@@ -2,7 +2,7 @@ package bio.terra.rbs.service.resource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import bio.terra.rbs.db.ResourceType;
+import bio.terra.rbs.common.ResourceType;
 import bio.terra.rbs.service.resource.flight.GoogleProjectCreationFlight;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
